Modified github workflow of release-drafter
It would create a new draft in every PR merge, but would also publish it
on every new version tag push. Since the dev. workflow is to create a PR
to bump the package version, this would always lead to a new draft being
created (with bad, stock formatting) and immediately being released,
without a chance to edit it.